# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Agroforestree-Sympla Integration API
  description: |
    Middleware para integração entre plataforma Sympla e API Agroforestree.
    
    Este serviço processa webhooks da Sympla e automaticamente cria doações 
    na plataforma Agroforestree quando pedidos são aprovados.
    
    ## Recursos Principais
    - Recebimento seguro de webhooks da Sympla
    - Validação de assinaturas HMAC
    - Integração com API Agroforestree
    - Dashboard de monitoramento de doações
    
    ## Segurança
    - Validação de assinaturas HMAC SHA-256
    - Tokens JWT para autenticação
    - Prevenção de ataques de replay
    
  version: 1.0.0
  contact:
    name: Equipe Agroforestree
    email: tech@agroforestree.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://agroforestree-sympla-middleware.onrender.com
    description: Servidor de produção (Render)
  - url: http://localhost:3001
    description: Servidor de desenvolvimento
paths:
  /webhooks/sympla:
    post:
      summary: Webhook da Sympla
      description: |
        Endpoint para receber webhooks da plataforma Sympla.
        
        ## Processo
        1. Valida assinatura HMAC do webhook
        2. Processa apenas eventos `order.approved`
        3. Cria doação automaticamente na API Agroforestree
        4. Armazena registro da tentativa de doação
        
        ## Segurança
        - Requer header `X-Sympla-Signature` com assinatura HMAC SHA-256
        - Validação de timestamp para prevenir replay attacks
        
      tags:
        - Webhooks
      security:
        - SymplaSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SymplaWebhookPayload'
            examples:
              order_approved:
                summary: Pedido Aprovado
                value:
                  event: "order.approved"
                  timestamp: "2024-01-01T10:00:00Z"
                  data:
                    order_identifier: "SPL-12345"
                    event_id: "EVT-98765"
                    event_name: "Workshop de Sustentabilidade"
                    total_order_amount: 120.00
                    buyer_first_name: "João"
                    buyer_last_name: "Silva"
                    buyer_email: "joao.silva@example.com"
                    order_status: "approved"
      responses:
        '200':
          description: Webhook processado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
              example:
                status: "success"
                message: "Webhook processado com sucesso"
                timestamp: "2024-01-01T10:00:00Z"
        '401':
          description: Assinatura inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid signature"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
...
  /api/donations:
    get:
      summary: Listar Doações
      description: |
        Retorna lista de todas as tentativas de doação processadas.
        
        ## Informações Incluídas
        - Status da doação (PENDING_USER_ACTION, CONSENT_GIVEN, COMPLETED, DECLINED)
        - Identificadores do pedido Sympla
        - Tokens de autenticação
        - Timestamps de criação e atualização
        
      tags:
        - Doações
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Lista de doações
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationsResponse'
              example:
                donations:
                  - id: "uuid-123"
                    sympla_order_id: "SPL-12345"
                    sympla_event_id: "EVT-98765"
                    status: "COMPLETED"
                    donation_token: "eyJhbGciOiJIUzI1NiIs..."
                    created_at: "2024-01-01T10:00:00Z"
                    updated_at: "2024-01-01T10:01:00Z"
                total: 1
                timestamp: "2024-01-01T10:00:00Z"
  /donation/initiate:
    get:
      summary: Página de Consentimento para Doação
      description: |
        Página que o usuário acessa após comprar ingresso (via link/QR Code).
        
        ## Fluxo LGPD-Compliant:
        1. Usuário compra ingresso na Sympla
        2. Recebe link/QR Code com donation_token
        3. Acessa esta página para dar consentimento explícito
        4. Escolhe se quer plantar uma árvore ou não
        
        ## Interface:
        - Informações do evento e impacto ambiental
        - Checkbox de consentimento explícito (LGPD)
        - Botões "Sim, plantar minha árvore" / "Não, obrigado"
        
      tags:
        - Fluxo do Usuário
      parameters:
        - name: token
          in: query
          required: true
          description: Token JWT da tentativa de doação
          schema:
            type: string
            example: "eyJhbGciOiJIUzI1NiIs..."
      responses:
        '200':
          description: Página de consentimento renderizada
          content:
            text/html:
              schema:
                type: string
              example: |
                <!DOCTYPE html>
                <html>
                  <head><title>Plante uma Árvore - Agroforestree</title></head>
                  <body>
                    <h1>🌱 Obrigado por apoiar eventos sustentáveis!</h1>
                    <p>Seu ingresso para <strong>Workshop Sustentabilidade</strong> foi confirmado.</p>
                    <form action="/donation/execute" method="POST">
                      <input type="hidden" name="token" value="...">
                      <label>
                        <input type="checkbox" name="consent" required>
                        Autorizo o plantio de uma árvore em meu nome (LGPD)
                      </label>
                      <button type="submit">🌳 Sim, plantar minha árvore!</button>
                    </form>
                  </body>
                </html>
        '400':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Token inválido ou expirado"
  /donation/execute:
    post:
      summary: Executar Doação com Consentimento
      description: |
        Processa o consentimento do usuário e executa a doação na Agroforestree.
        
        ## Processo:
        1. Valida o token JWT da tentativa
        2. Verifica consentimento explícito
        3. Chama API Agroforestree para criar doação
        4. Redireciona para página de confirmação com certificado
        
        ## Conformidade LGPD:
        - Consentimento explícito e informado
        - Finalidade específica (plantio de árvore)
        - Possibilidade de recusa sem prejuízo
        
      tags:
        - Fluxo do Usuário
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
                - consent
              properties:
                token:
                  type: string
                  description: Token JWT da tentativa de doação
                  example: "eyJhbGciOiJIUzI1NiIs..."
                consent:
                  type: string
                  enum: ['on']
                  description: Checkbox de consentimento (deve ser 'on')
      responses:
        '302':
          description: Redirecionamento para página de confirmação
          headers:
            Location:
              schema:
                type: string
                example: "/donation/success?donationId=AGF-1234567890"
        '400':
          description: Dados inválidos ou consentimento não dado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Consentimento obrigatório para prosseguir"
  /donation/success:
    get:
      summary: Página de Sucesso da Doação
      description: |
        Página de confirmação após doação bem-sucedida.
        
        ## Conteúdo:
        - Confirmação da doação realizada
        - Link para certificado PDF
        - Informações sobre impacto ambiental
        - Compartilhamento social
        
      tags:
        - Fluxo do Usuário
      parameters:
        - name: donationId
          in: query
          required: true
          description: ID da doação criada na Agroforestree
          schema:
            type: string
            example: "AGF-1234567890"
      responses:
        '200':
          description: Página de confirmação renderizada
          content:
            text/html:
              schema:
                type: string
  /events/{eventId}/impact-summary:
    get:
      summary: Resumo de Impacto por Evento
      description: |
        Dashboard executivo para organizadores de evento.
        
        ## Métricas Incluídas:
        - Total de doações/árvores plantadas
        - Taxa de engajamento (% de compradores que doaram)
        - Impacto de CO₂ compensado
        - Agricultores familiares beneficiados
        
        ## Caso de Uso:
        - Relatórios pós-evento para organizadores
        - Métricas ESG para relatórios corporativos
        - Dados para marketing de próximos eventos
        
      tags:
        - Dashboard Organizador
      security:
        - ApiKeyAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID do evento na Sympla
          schema:
            type: string
            example: "EVT-98765"
      responses:
        '200':
          description: Métricas de impacto do evento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventImpactSummary'
              example:
                eventId: "EVT-98765"
                eventName: "Workshop de Sustentabilidade"
                summary:
                  totalTickets: 600
                  totalDonations: 152
                  engagementRate: "25.3%"
                  treesPlanted: 152
                  co2Compensated: "25.08 toneladas"
                  farmersSupported: 152
                  totalDonationValue: "R$ 760,00"
                period:
                  startDate: "2024-01-01T00:00:00Z"
                  endDate: "2024-01-01T23:59:59Z"
        '401':
          description: Chave de API inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Evento não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{eventId}/qrcode:
    get:
      summary: Gerar QR Code para Evento
      description: |
        Gera QR Code personalizado para o evento, apontando para página de doação.
        
        ## Funcionalidades:
        - QR Code dinâmico com logo da Agroforestree
        - Link direto para /donation/initiate
        - Personalizável com cores do evento
        - Diferentes tamanhos (query parameter)
        
        ## Uso:
        - Telões durante o evento
        - Material gráfico de divulgação
        - Redes sociais do organizador
        - Email marketing pós-evento
        
      tags:
        - Dashboard Organizador
      security:
        - ApiKeyAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID do evento na Sympla
          schema:
            type: string
            example: "EVT-98765"
        - name: size
          in: query
          required: false
          description: Tamanho do QR Code em pixels
          schema:
            type: integer
            enum: [200, 400, 800]
            default: 400
        - name: format
          in: query
          required: false
          description: Formato da imagem
          schema:
            type: string
            enum: ['png', 'svg']
            default: 'png'
      responses:
        '200':
          description: Imagem do QR Code gerada
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/svg+xml:
              schema:
                type: string
        '401':
          description: Chave de API inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Evento não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /simulate/sympla-webhook:
    post:
      summary: Simula envio de webhook da Sympla para testes
      description: |
        Endpoint auxiliar para simular localmente o envio de um webhook da Sympla, 
        gerando assinatura HMAC válida e testando o fluxo de doação.

        Esta rota envia um `order.approved` artificial para o endpoint oficial `/webhooks/sympla`.
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventType:
                  type: string
                  enum: ['order.approved', 'order.cancelled', 'order.refunded']
                  default: 'order.approved'
                  description: Tipo de evento a ser simulado
                order_identifier:
                  type: string
                  example: XYZ123
                event_id:
                  type: string
                  example: EVT456
                event_name:
                  type: string
                  example: Festival Sustentável 2025
                buyer_first_name:
                  type: string
                  example: João
                buyer_last_name:
                  type: string
                  example: Silva
                buyer_email:
                  type: string
                  format: email
                  example: joao@email.com
                total_order_amount:
                  type: number
                  example: 50.0
      responses:
        '200':
          description: Webhook simulado enviado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ✅ Simulação enviada
                  webhookResponse:
                    type: object
        '500':
          description: Erro ao simular webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    SymplaSignature:
      type: apiKey
      in: header
      name: X-Sympla-Signature
      description: |
        Assinatura HMAC SHA-256 do corpo da requisição.
        
        Formato: `sha256={hash}`
        
        O hash é calculado usando HMAC SHA-256 com a chave secreta configurada
        no webhook da Sympla.

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Chave de API para acesso às rotas de gestão e dashboard.
        
        Esta chave protege endpoints sensíveis como:
        - Listagem de doações
        - Métricas de eventos
        - Geração de QR Codes
        
        Obtenha sua chave no painel administrativo da Agroforestree.
  schemas:
    SymplaWebhookPayload:
      type: object
      required:
        - event
        - data
      properties:
        event:
          type: string
          enum: ['order.approved', 'order.created', 'order.cancelled', 'order.refunded']
          description: Tipo do evento
        data:
          $ref: '#/components/schemas/SymplaOrderData'
        timestamp:
          type: string
          format: date-time
          description: Timestamp do evento
...
    DonationAttempt:
      type: object
      properties:
        id:
          type: string
          description: ID único da tentativa
          example: "uuid-123"
        sympla_order_id:
          type: string
          description: ID do pedido Sympla
          example: "SPL-12345"
        sympla_event_id:
          type: string
          description: ID do evento Sympla
          example: "EVT-98765"
        status:
          type: string
          enum: ['PENDING_USER_ACTION', 'CONSENT_GIVEN', 'COMPLETED', 'DECLINED']
          description: Status da doação
        donation_token:
          type: string
          description: Token JWT para a doação
          example: "eyJhbGciOiJIUzI1NiIs..."
        created_at:
          type: string
          format: date-time
          description: Data de criação
        updated_at:
          type: string
          format: date-time
          description: Data de atualização
    WebhookResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Webhook processado com sucesso"
        timestamp:
          type: string
          format: date-time
...
    DonationsResponse:
      type: object
      properties:
        donations:
          type: array
          items:
            $ref: '#/components/schemas/DonationAttempt'
        total:
          type: integer
          description: Total de doações
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: "Invalid signature"
    EventImpactSummary:
      type: object
      description: Resumo de impacto ambiental de um evento específico
      properties:
        eventId:
          type: string
          description: ID do evento na Sympla
          example: "EVT-98765"
        eventName:
          type: string
          description: Nome do evento
          example: "Workshop de Sustentabilidade"
        summary:
          type: object
          description: Métricas consolidadas do evento
          properties:
            totalTickets:
              type: integer
              description: Total de ingressos vendidos
              example: 600
            totalDonations:
              type: integer
              description: Total de doações/árvores plantadas
              example: 152
            engagementRate:
              type: string
              description: Taxa de engajamento (% de compradores que doaram)
              example: "25.3%"
            treesPlanted:
              type: integer
              description: Número de árvores plantadas
              example: 152
            co2Compensated:
              type: string
              description: CO₂ compensado em toneladas
              example: "25.08 toneladas"
            farmersSupported:
              type: integer
              description: Agricultores familiares beneficiados
              example: 152
            totalDonationValue:
              type: string
              description: Valor total das doações
              example: "R$ 760,00"
        period:
          type: object
          description: Período de análise
          properties:
            startDate:
              type: string
              format: date-time
              description: Data de início do evento
              example: "2024-01-01T00:00:00Z"
            endDate:
              type: string
              format: date-time
              description: Data de fim do evento
              example: "2024-01-01T23:59:59Z"
...